Refactor: Mongo DB Document를 각 도메인에 알맞은 디렉토리로 이동
