+++ src/main/java/com/team7/retriever/domain/agrot/controller/ArgotsController.java
@@ -1,7 +1,7 @@
 package com.team7.retriever.domain.agrot.controller;
 
 
-import com.team7.retriever.entity.Argot;
+import com.team7.retriever.domain.agrot.domain.document.Argot;
 import com.team7.retriever.service.ArgotsService;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
