+++ src/main/java/com/team7/retriever/domain/agrot/domain/document/Argot.java
@@ -1,12 +1,10 @@
-package com.team7.retriever.entity;
+package com.team7.retriever.domain.agrot.domain.document;
 
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.Field;
-
-import java.time.LocalDateTime;
 
 @Getter
 @Setter
