+++ src/main/java/com/team7/retriever/domain/bookmark/controller/BookmarksController.java
@@ -1,7 +1,7 @@
 package com.team7.retriever.domain.bookmark.controller;
 
 import com.team7.retriever.domain.bookmark.controller.dto.request.BookmarksRequest;
-import com.team7.retriever.entity.Bookmarks;
+import com.team7.retriever.domain.bookmark.domain.document.Bookmarks;
 import com.team7.retriever.service.BookmarksService;
 
 import lombok.AllArgsConstructor;
