+++ src/main/java/com/team7/retriever/domain/channel/controller/ChSimilarityController.java
@@ -1,6 +1,6 @@
 package com.team7.retriever.domain.channel.controller;
 
-import com.team7.retriever.entity.ChannelSimilarity;
+import com.team7.retriever.domain.channel.domain.document.ChannelSimilarity;
 import com.team7.retriever.service.ChSimilarityService;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
