+++ src/main/java/com/team7/retriever/domain/channel/controller/ChannelDataController.java
@@ -1,6 +1,6 @@
 package com.team7.retriever.domain.channel.controller;
 
-import com.team7.retriever.entity.ChData;
+import com.team7.retriever.domain.channel.domain.document.ChData;
 import com.team7.retriever.service.ChDataService;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
