+++ src/main/java/com/team7/retriever/domain/channel/domain/document/ChData.java
@@ -1,4 +1,4 @@
-package com.team7.retriever.entity;
+package com.team7.retriever.domain.channel.domain.document;
 
 import lombok.Getter;
 import org.springframework.data.annotation.Id;
