+++ src/main/java/com/team7/retriever/domain/chatbot/domain/document/ChatBots.java
@@ -1,11 +1,10 @@
-package com.team7.retriever.entity;
+package com.team7.retriever.domain.chatbot.domain.document;
 
 import lombok.Getter;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 
 @Getter
