+++ src/main/java/com/team7/retriever/repository/ArgotsRepository.java
@@ -1,6 +1,6 @@
 package com.team7.retriever.repository;
 
-import com.team7.retriever.entity.Argot;
+import com.team7.retriever.domain.agrot.domain.document.Argot;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.mongodb.repository.Query;
