+++ src/main/java/com/team7/retriever/repository/ChannelSimilarityRepository.java
@@ -1,6 +1,6 @@
 package com.team7.retriever.repository;
 
-import com.team7.retriever.entity.ChannelSimilarity;
+import com.team7.retriever.domain.channel.domain.document.ChannelSimilarity;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
