+++ src/main/java/com/team7/retriever/repository/ChatBotsRepository.java
@@ -1,6 +1,6 @@
 package com.team7.retriever.repository;
 
-import com.team7.retriever.entity.ChatBots;
+import com.team7.retriever.domain.chatbot.domain.document.ChatBots;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
