+++ src/main/java/com/team7/retriever/repository/DrugsRepository.java
@@ -1,6 +1,6 @@
 package com.team7.retriever.repository;
 
-import com.team7.retriever.entity.Drugs;
+import com.team7.retriever.domain.drug.domain.document.Drugs;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.mongodb.repository.Query;
