+++ src/main/java/com/team7/retriever/repository/PostHtmlRepository.java
@@ -1,6 +1,6 @@
 package com.team7.retriever.repository;
 
-import com.team7.retriever.entity.PostHtml;
+import com.team7.retriever.domain.crawling.domain.document.PostHtml;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface PostHtmlRepository extends MongoRepository<PostHtml, String> {
