+++ src/main/java/com/team7/retriever/service/BookmarksService.java
@@ -1,6 +1,6 @@
 package com.team7.retriever.service;
 
-import com.team7.retriever.entity.Bookmarks;
+import com.team7.retriever.domain.bookmark.domain.document.Bookmarks;
 import com.team7.retriever.repository.BookmarksRepository;
 
 import org.springframework.stereotype.Service;
