+++ src/main/java/com/team7/retriever/service/ChDataService.java
@@ -1,6 +1,6 @@
 package com.team7.retriever.service;
 
-import com.team7.retriever.entity.ChData;
+import com.team7.retriever.domain.channel.domain.document.ChData;
 import com.team7.retriever.repository.ChDataRepository;
 
 import org.springframework.stereotype.Service;
