+++ src/main/java/com/team7/retriever/service/ChatBotsService.java
@@ -1,6 +1,6 @@
 package com.team7.retriever.service;
 
-import com.team7.retriever.entity.ChatBots;
+import com.team7.retriever.domain.chatbot.domain.document.ChatBots;
 import com.team7.retriever.repository.ChatBotsRepository;
 
 import org.springframework.stereotype.Service;
