+++ src/main/java/com/team7/retriever/service/DrugsService.java
@@ -1,6 +1,6 @@
 package com.team7.retriever.service;
 
-import com.team7.retriever.entity.Drugs;
+import com.team7.retriever.domain.drug.domain.document.Drugs;
 import com.team7.retriever.repository.DrugsRepository;
 
 import org.springframework.stereotype.Service;
