+++ src/main/java/com/team7/retriever/service/PreprocessService.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.team7.retriever.domain.crawling.controller.dto.response.PreprocessResponse;
-import com.team7.retriever.entity.Posts;
+import com.team7.retriever.domain.post.domain.document.Posts;
 import com.team7.retriever.repository.PostsRepository;
 
 import lombok.RequiredArgsConstructor;
